change duration in paper (int to float)

+++ src/Core/Application/Examination/Papers/Dtos/PaperDeletedDto.cs
@@ -6,7 +6,7 @@ public class PaperDeletedDto : IDto
     public Guid Id { get; set; }
     public string ExamName { get; set; }
     public SubjectDto Subject { get; set; }
-    public int Duration { get; set; }
+    public float Duration { get; set; }
     public DateTime CreatedOn { get; set; }
     public DateTime? DeletedOn { get; set; }
 }
